feat(store): sql-only-store
- add getPage(...) to sql-message-store
- acts as separate store impl that can be activated with
  `--sqlite-only-store`

+++ waku/v2/node/storage/message/waku_message_store.nim
@@ -8,6 +8,7 @@ import
   ./message_store,
   ../sqlite,
   ../../../protocol/waku_message,
+  ../../../protocol/waku_store/waku_store,
   ../../../utils/pagination,
   ../../../utils/time
 
@@ -37,6 +38,7 @@ type
     storeCapacity: int # represents both the number of messages that are persisted in the sqlite DB (excl. the overflow window explained above), and the number of messages that get loaded via `getAll`.
     storeMaxLoad: int  # = storeCapacity * MaxStoreOverflow
     deleteWindow: int  # = (storeCapacity * MaxStoreOverflow - storeCapacity)/2; half of the overflow window, the amount of messages deleted when overflow occurs
+    isSqliteOnly: bool 
  
 proc messageCount(db: SqliteDatabase): MessageStoreResult[int64] =
   var numMessages: int64
@@ -71,7 +73,7 @@ proc deleteOldest(db: WakuMessageStore): MessageStoreResult[void] =
 
   ok()
 
-proc init*(T: type WakuMessageStore, db: SqliteDatabase, storeCapacity: int = 50000): MessageStoreResult[T] =
+proc init*(T: type WakuMessageStore, db: SqliteDatabase, storeCapacity: int = 50000, isSqliteOnly = false): MessageStoreResult[T] =
   ## Table is the SQL query for creating the messages Table.
   ## It contains:
   ##  - 4-Byte ContentTopic stored as an Integer
@@ -112,10 +114,11 @@ proc init*(T: type WakuMessageStore, db: SqliteDatabase, storeCapacity: int = 50
                       numMessages: int(numMessages),
                       storeCapacity: storeCapacity,
                       storeMaxLoad: storeMaxLoad,
-                      deleteWindow: deleteWindow)
+                      deleteWindow: deleteWindow,
+                      isSqliteOnly: isSqliteOnly)
 
-  # If the loaded db is already over max load, delete the oldest messages before returning the WakuMessageStore object
-  if wms.numMessages >= wms.storeMaxLoad:
+  # if the in-memory store is used and if the loaded db is already over max load, delete the oldest messages before returning the WakuMessageStore object
+  if not isSqliteOnly and wms.numMessages >= wms.storeMaxLoad:
     let res = wms.deleteOldest()
     if res.isErr: return err("deleting oldest messages failed")
 
@@ -146,13 +149,12 @@ method put*(db: WakuMessageStore, cursor: Index, message: WakuMessage, pubsubTop
     return err("failed")
 
   db.numMessages += 1
-  if db.numMessages >= db.storeMaxLoad:
+  # if the in-memory store is used and if the loaded db is already over max load, delete the oldest messages
+  if not db.isSqliteOnly and db.numMessages >= db.storeMaxLoad:
     let res = db.deleteOldest()
     if res.isErr: return err("deleting oldest failed")
 
   ok()
-
-
 
 method getAll*(db: WakuMessageStore, onData: message_store.DataProc): MessageStoreResult[bool] =
   ## Retrieves `storeCapacity` many  messages from the storage.
@@ -207,6 +209,145 @@ method getAll*(db: WakuMessageStore, onData: message_store.DataProc): MessageSto
     return err(res.error)
 
   ok gotMessages
+
+
+method getPage*(db: WakuMessageStore,
+              pred: QueryFilterMatcher,
+              pagingInfo: PagingInfo):
+             (seq[WakuMessage], PagingInfo, HistoryResponseError) =
+  ## Get a single page of history matching the predicate and
+  ## adhering to the pagingInfo parameters
+  
+  trace "getting page from SQLite DB", pagingInfo=pagingInfo
+
+  let
+    maxPageSize = if pagingInfo.pageSize == 0 or pagingInfo.pageSize > MaxPageSize: MaxPageSize # Used default MaxPageSize for invalid pagingInfos
+                  else: pagingInfo.pageSize
+
+  var cursor = pagingInfo.cursor 
+
+  var messages: seq[WakuMessage]
+  var
+    lastIndex: Index
+    numRecordsVisitedPage: uint64 = 0  # number of DB records visited during retrieving the last page from the DB
+    numRecordsVisitedTotal: uint64 = 0 # number of DB records visited in total
+
+  proc msg(s: ptr sqlite3_stmt) = # this is the actual onData proc that is passed to the query proc (the message store adds one indirection)
+    let
+      receiverTimestamp = column_timestamp(s, 0)
+
+      topic = cast[ptr UncheckedArray[byte]](sqlite3_column_blob(s, 1))
+      topicLength = sqlite3_column_bytes(s,1)
+      contentTopic = ContentTopic(string.fromBytes(@(toOpenArray(topic, 0, topicLength-1))))
+
+      p = cast[ptr UncheckedArray[byte]](sqlite3_column_blob(s, 2))
+      length = sqlite3_column_bytes(s, 2)
+      payload = @(toOpenArray(p, 0, length-1))
+
+      pubsubTopicPointer = cast[ptr UncheckedArray[byte]](sqlite3_column_blob(s, 3))
+      pubsubTopicLength = sqlite3_column_bytes(s,3)
+      pubsubTopic = string.fromBytes(@(toOpenArray(pubsubTopicPointer, 0, pubsubTopicLength-1)))
+
+      version = sqlite3_column_int64(s, 4)
+
+      senderTimestamp = column_timestamp(s, 5)
+      retMsg = WakuMessage(contentTopic: contentTopic, payload: payload, version: uint32(version), timestamp: Timestamp(senderTimestamp))
+      # TODO: we should consolidate WakuMessage, Index, and IndexedWakuMessage; reason: avoid unnecessary copying and recalculation
+      # TODO: retrieve digest from DB
+      index = retMsg.computeIndex(receiverTimestamp, pubsubTopic)
+      indexedWakuMsg = IndexedWakuMessage(msg: retMsg, index: index, pubsubTopic: pubsubTopic)
+
+    lastIndex = index
+    numRecordsVisitedPage += 1
+    try:
+      if pred(indexedWakuMsg): #TODO throws unknown exception
+        messages.add(retMsg)
+    except:
+      # TODO properly handle this exception
+      quit 1
+
+  # TODO: deduplicate / condense the following 4 DB query strings
+  # If no index has been set in pagingInfo, start with the first message (or the last in case of backwards direction)
+  if cursor == Index(): ## TODO: pagingInfo.cursor should be an Option. We shouldn't rely on empty initialisation to determine if set or not!
+    let noCursorQuery = if pagingInfo.direction == PagingDirection.FORWARD:
+        "SELECT receiverTimestamp, contentTopic, payload, pubsubTopic, version, senderTimestamp " &
+        "FROM " & TABLE_TITLE & " " &
+        "ORDER BY senderTimestamp, receiverTimestamp, id, pubsubTopic " &
+        "LIMIT " & $maxPageSize & ";"
+    else:
+        "SELECT receiverTimestamp, contentTopic, payload, pubsubTopic, version, senderTimestamp " &
+        "FROM " & TABLE_TITLE & " " &
+        "ORDER BY senderTimestamp DESC, receiverTimestamp DESC, id DESC, pubsubTopic DESC " &
+        "LIMIT " & $maxPageSize & ";"
+
+    let res = db.database.query(noCursorQuery, msg)
+    if res.isErr:
+      # TODO properly handle this exception
+      quit 1
+    numRecordsVisitedTotal = numRecordsVisitedPage
+    numRecordsVisitedPage = 0
+    cursor = lastIndex
+
+  let preparedPageQuery = if pagingInfo.direction == PagingDirection.FORWARD:
+      db.database.prepareStmt(
+        "SELECT receiverTimestamp, contentTopic, payload, pubsubTopic, version, senderTimestamp " &
+        "FROM " & TABLE_TITLE & " " &
+        "WHERE (senderTimestamp, id, pubsubTopic) > (?, ?, ?)" &
+        "ORDER BY senderTimestamp, id, receiverTimestamp, pubsubTopic " &
+        "LIMIT " & $maxPageSize & ";",
+        (Timestamp, seq[byte], seq[byte]),
+        (Timestamp, seq[byte], seq[byte], seq[byte], int64, Timestamp),
+      )
+    else:
+      db.database.prepareStmt(
+        "SELECT receiverTimestamp, contentTopic, payload, pubsubTopic, version, senderTimestamp " &
+        "FROM " & TABLE_TITLE & " " &
+        "WHERE (senderTimestamp, id, pubsubTopic) < (?, ?, ?)" &
+        "ORDER BY senderTimestamp DESC, receiverTimestamp DESC, id DESC, pubsubTopic DESC " &
+        "LIMIT " & $maxPageSize & ";",
+        (Timestamp, seq[byte], seq[byte]),
+        (Timestamp, seq[byte], seq[byte], seq[byte], int64, Timestamp),
+      )
+
+  if preparedPageQuery.isErr:
+    # return err("failed to prepare page query")
+    quit 1 # TODO
+
+  # TODO: DoS attack migration against: sending a lot of queries with sparse (or non-matching) filters making the store node run through the whole DB. Even worse with pageSize = 1.
+  while uint64(messages.len) < maxPageSize:
+    let res = preparedPageQuery.value.exec((cursor.senderTime, @(cursor.digest.data), cursor.pubsubTopic.toBytes()), msg)
+    if res.isErr:
+      # return err("failed")
+      quit 1 # TODO
+    numRecordsVisitedTotal += numRecordsVisitedPage
+    if numRecordsVisitedPage == 0: break # we are at the end of the DB (find more efficient/integrated solution to track that event)
+    numRecordsVisitedPage = 0
+    cursor = lastIndex
+
+  let outPagingInfo = PagingInfo(pageSize: messages.len.uint,
+                             cursor: lastIndex,
+                             direction: pagingInfo.direction)
+
+  let historyResponseError = if numRecordsVisitedTotal == 0: HistoryResponseError.INVALID_CURSOR # Index is not in DB (also if queried Index points to last entry)
+    else: HistoryResponseError.NONE
+
+  return (messages, outPagingInfo, historyResponseError) # TODO: use canonical way of handling errors
+
+
+
+
+method getPage*(db: WakuMessageStore,
+              pagingInfo: PagingInfo):
+             (seq[WakuMessage], PagingInfo, HistoryResponseError) =
+  ## Get a single page of history without filtering.
+  ## Adhere to the pagingInfo parameters
+  
+  proc predicate(i: IndexedWakuMessage): bool = true # no filtering
+
+  return getPage(db, predicate, pagingInfo)
+
+
+
 
 proc close*(db: WakuMessageStore) = 
   ## Closes the database.
